backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
xgboost==2.0.3
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.2
httpx==0.25.2
earthengine-api==0.1.384
firebase-admin==6.2.0
joblib==1.3.2
python-multipart==0.0.6
google-generativeai==0.8.3
